+++ docs/README.md
@@ -17,7 +17,7 @@
 
 - [x] 자동차들이 무작위로 전진한다. 
   - 전진하는 조건은 0에서 9 사이에서 무작위 값을 구한 후 무작위 값이 4 이상일 경우이다.
-- [ ] 자동차들이 전진한 결과를 화면에 출력한다. 
+- [x] 자동차들이 전진한 결과를 화면에 출력한다. 
 - [ ] 우승자를 판별한다. 
 - [ ] 우승자 결과를 화면에 출력한다. (우승자가 여러 명일 경우 쉼표(,)를 이용하여 구분한다.)
 - [ ] 다시 시작하기 버튼을 클릭한다.
