+++ src/model/getWinners.js
@@ -1,0 +1,3 @@
+function getWinners() {
+  
+}
